authenticator/requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
protobuf==3.20.0
PyJWT==2.3.0
pycryptodome==3.14.1
pymongo==4.1.1
